IES_2425_113372-master/Lab05/Lab05_2/poetry-demo/pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["brunotavaresz <brunotavaresmeixedo@ua.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"